+++ example/runtime/import.js
@@ -1,4 +1,4 @@
-import { ECIES_CONFIG, utils as _utils } from "eciesjs";
+import { ECIES_CONFIG } from "eciesjs";
 import * as config from "eciesjs/config";
 import * as consts from "eciesjs/consts";
 import * as utils from "eciesjs/utils";
@@ -7,4 +7,3 @@ console.log("ECIES_CONFIG:", ECIES_CONFIG);
 console.log("config:", config);
 console.log("consts:", consts);
 console.log("utils:", utils);
-console.log("index utils:", _utils);