+++ src/types.ts
@@ -1,2 +1,0 @@
-export const IS_BUFFER_SUPPORTED =
-  typeof Buffer !== "undefined" && typeof Buffer.from === "function";